# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MSPR HealthAI Coach - AI Nutrition",
    "description": "\nMicro-service d'analyse nutritionnelle et de generation de plans repas par IA, partie de la plateforme MSPR HealthAI Coach.\n\nDeux pipelines d'IA sont exposes :\n\n- **Analyse de repas** : photo en entree, classification HuggingFace (`nateraw/food`, modele Food-101), lookup nutritionnel sur les datasets ETL, detection de desequilibres macros vs profil utilisateur, puis recommandations generees par Ollama (Gemma3:4b) avec fallback matrice statique.\n- **Generation de plans repas** : objectif de sante, preferences alimentaires et budget, prompt structure soumis a Ollama, validation du JSON et persistance du plan sur 1 a 30 jours.\n\nLe service expose egalement la gestion du profil nutritionnel (`nutrition-goals/me`) et l'historique pagine des analyses et des plans (`meal-analyses/me`, `meal-plans/me`). L'authentification se fait via JWT signe par MSPR-AUTH (`Authorization: Bearer <jwt>`).\n",
    "contact": {
      "name": "Arthur Poncin",
      "email": "arthur.poncin@sixense-group.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/me/macros": {
      "get": {
        "tags": [
          "Profil"
        ],
        "summary": "Cible journaliere TDEE et macros pour l'utilisateur",
        "description": "Retourne la cible journaliere de l'utilisateur authentifie (TDEE et macros).\n\nLe TDEE est calcule via la formule de Mifflin-St Jeor a partir des champs\nbiometriques du profil (`gender`, `age`, `weight_kg`, `height_cm`, `activity_level`)\net de l'objectif sante (`health_goal`). La repartition macro suit l'objectif :\n\n- `balance` : 50 % glucides, 20 % proteines, 30 % lipides\n- `weight_loss` : 40 / 30 / 30\n- `muscle_gain` : 45 / 30 / 25\n- `sport_performance` : 55 / 20 / 25\n\nQuand un champ biometrique est manquant, la reponse renseigne\n`profile_completion_required=true` et liste les `missing_fields` plutot que de\nretourner une cible degradee.\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis.",
        "operationId": "get_my_macros_api_v1_me_macros_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cible journaliere ou liste des champs manquants.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeMacrosResponse"
                },
                "examples": {
                  "complete_profile": {
                    "summary": "Profil complet : TDEE et macros calcules",
                    "value": {
                      "profile_completion_required": false,
                      "missing_fields": [],
                      "tdee": 2450,
                      "macros": {
                        "calories": 2450,
                        "protein_g": 184.0,
                        "carbs_g": 276.0,
                        "fat_g": 68.0
                      }
                    }
                  },
                  "incomplete_profile": {
                    "summary": "Profil incomplet : champs manquants",
                    "value": {
                      "profile_completion_required": true,
                      "missing_fields": [
                        "weight_kg",
                        "height_cm"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Validation Pydantic en echec (type ou contrainte de champ violee)."
          },
          "503": {
            "description": "Service degrade : dependance externe (Ollama) injoignable et fallback indisponible."
          }
        }
      }
    },
    "/api/v1/me/preferences": {
      "get": {
        "tags": [
          "Profil"
        ],
        "summary": "Preferences LLM de l'utilisateur",
        "operationId": "get_my_preferences_api_v1_me_preferences_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Profil"
        ],
        "summary": "Met a jour les preferences LLM de l'utilisateur",
        "operationId": "patch_my_preferences_api_v1_me_preferences_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMBackend": {
        "type": "string",
        "enum": [
          "ollama",
          "mistral"
        ],
        "title": "LLMBackend",
        "description": "Backends LLM supportes (PRD #71). Toute autre valeur est rejetee."
      },
      "MacroTargetsView": {
        "properties": {
          "calories": {
            "type": "number",
            "title": "Calories"
          },
          "protein_g": {
            "type": "number",
            "title": "Protein G"
          },
          "carbs_g": {
            "type": "number",
            "title": "Carbs G"
          },
          "fat_g": {
            "type": "number",
            "title": "Fat G"
          }
        },
        "type": "object",
        "required": [
          "calories",
          "protein_g",
          "carbs_g",
          "fat_g"
        ],
        "title": "MacroTargetsView"
      },
      "MeMacrosResponse": {
        "properties": {
          "profile_completion_required": {
            "type": "boolean",
            "title": "Profile Completion Required"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields",
            "default": []
          },
          "tdee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tdee"
          },
          "macros": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MacroTargetsView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "profile_completion_required"
        ],
        "title": "MeMacrosResponse"
      },
...
      "PreferencesUpdate": {
        "properties": {
          "preferred_llm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMBackend"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "PreferencesUpdate",
        "description": "Payload PATCH /me/preferences. preferred_llm=null reset au defaut env."
      },
      "PreferencesView": {
        "properties": {
          "preferred_llm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMBackend"
              },
              {
                "type": "null"
              }
            ]
          },
          "effective_llm": {
            "$ref": "#/components/schemas/LLMBackend"
          }
        },
        "type": "object",
        "required": [
          "effective_llm"
        ],
        "title": "PreferencesView",
        "description": "Reponse GET/PATCH /me/preferences.\n\npreferred_llm : valeur brute persistee pour cet utilisateur (None si jamais\n                definie ou explicitement reset).\neffective_llm : backend effectivement utilise par l'orchestrator\n                (preference user OR settings.default_llm)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}